requirements.txt
asgiref==3.3.4
biplist==1.0.3
defusedxml==0.7.1
Django==3.2
django-rest-framework-mongoengine==3.4.1
django-tastypie==0.14.3
django-tastypie-mongoengine==0.4.6
djangorestframework==3.12.4
lxml==4.6.3
mongoengine==0.8.1
pymongo==2.8
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
